Refactor all tests and production code that uses Score, AverageRating, and TotalRatingAccumulated:     * From using the double primitive type to create a BigDecimal to       using a string. Using the double to create an instance of a BigDecimal is inaccurate

+++ src/main/java/com/example/demo/domain/averageCalculator/Calculator.java
@@ -1,7 +1,9 @@
 package com.example.demo.domain.averageCalculator;
 
+import java.math.BigDecimal;
+
 public interface Calculator {
 
-    double calculateAverage();
+    BigDecimal calculateAverage();
 
 }
